Added the .npmrc file to the root level. Updated the Client folder to reflect the change from CRA to Vite
