+++ Develop/.npmrc
@@ -1,0 +1,3 @@
+production=false
+
+progress=false
