+++ Develop/client/src/App.jsx
@@ -1,0 +1,15 @@
+import './App.css';
+import { Outlet } from 'react-router-dom';
+
+import Navbar from './components/Navbar';
+
+function App() {
+  return (
+    <>
+      <Navbar />
+      <Outlet />
+    </>
+  );
+}
+
+export default App;
