+++ Develop/server/package.json
@@ -11,10 +11,9 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "bcrypt": "^5.0.0",
+    "bcrypt": "^4.0.1",
     "express": "^4.17.1",
-    "graphql": "^15.5.0",
-    "jsonwebtoken": "^9.0.0",
+    "jsonwebtoken": "^8.5.1",
     "mongoose": "^7.0.2"
   },
   "devDependencies": {
